`StypOptions.addSheet` is required now
